python-old/requirements.txt
rich
pyperclip
fastapi
PyYAML
gunicorn
uvicorn[standard]
dotenv